feat(newOnboarding): adds content on step3
