+++ src/components/atoms/Paragraph/Paragraph.ts
@@ -1,9 +1,14 @@
 import styled from 'styled-components';
 
-const Paragraph = styled.p`
+interface ParagraphProps {
+  align?: 'right' | 'left' | 'center';
+}
+
+const Paragraph = styled.p<ParagraphProps>`
   margin: 0;
   font-size: ${props => props.theme.fontSizeDefault};
   color: ${props => props.theme.textColor};
+  text-align: ${props => props.align || 'center'};
 
   @media (max-width: ${props => props.theme.tabletWidth}) {
     font-size: 18px;
